requirements.txt
canonicalwebteam.flask-base==0.3.3